"Spotlight Importer.xctemplate" from Xcode 8.2.1
I assume this template is freely redistributable, as any program which
includes a Spotlight Importer will necessarily have all of this source
code.  It even includes a ___COPYRIGHT___ template substitution
variable, so they clearly intend for the copyright (even for their
hundreds of lines of boilerplate) to be granted to any third-party Mac
developer.

(I'm also certainly not the first person to post this code publicly.)

This template was inexplicably removed from Xcode 9 (and subsequent
releases), making it impossible to write a new Spotlight Importer with
any Xcode since 2016.  (The Spotlight documentation does not contain
nearly enough information to reproduce all of the CFPlugin
interfaces.)  Otherwise, you'd have to download an old 4.5GB copy of
Xcode from Apple.com to write a Spotlight Importer -- which, as they
say, all applications with custom document formats should do.

To use this, copy it into:
    'Xcode.app/Contents/Developer/Library/Xcode/Templates/Project Templates/Mac/Other/'.